requirements.txt
# Core dependencies for zcore project
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
matplotlib>=3.5.0
fiftyone>=0.16.0
IPython>=7.30.0

# HuggingFace transformers for one-shot classification
transformers[torch]>=4.20.0
pillow>=8.0.0

# CLIP embeddings support
open-clip-torch>=2.7.0

# YOLO embeddings support
ultralytics>=8.0.0

# Development and utility dependencies
argparse  # Built-in but listed for clarity
pickle5>=1.0.7; python_version < "3.8"

# quality of life tools
pandas>=1.3.0
scikit-learn>=0.24.0
tqdm>=4.62.0
requests>=2.26.0
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0